Fix Product Reviews

+++ composeApp/src/commonMain/kotlin/domain/Models/Appointment.kt
@@ -9,9 +9,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable @Parcelize
 data class Appointment(
-    @SerialName("id") val appointmentId: Long? = -1, @SerialName("user_id") val userId: Int? = -1, @SerialName("vendor_id") val vendorId: Int = -1,
-    @SerialName("serviceLocation") val serviceLocation: String = ServiceLocationEnum.SPA.toPath(), @SerialName("serviceStatus") val serviceStatus: String = ServiceStatusEnum.PENDING.toPath(), @SerialName("service_id") var serviceId: Int = -1,
-    @SerialName("therapist_id") val therapistId: Int = -1, @SerialName("service_type_id") var serviceTypeId: Int? = -1, @SerialName("appointmentTime") var appointmentTime: Int? = -1,
+    @SerialName("id") val appointmentId: Long? = -1, @SerialName("user_id") val userId: Long? = -1, @SerialName("vendor_id") val vendorId: Long = -1,
+    @SerialName("serviceLocation") val serviceLocation: String = ServiceLocationEnum.SPA.toPath(), @SerialName("serviceStatus") val serviceStatus: String = ServiceStatusEnum.PENDING.toPath(), @SerialName("service_id") var serviceId: Long = -1,
+    @SerialName("therapist_id") val therapistId: Long = -1, @SerialName("service_type_id") var serviceTypeId: Long? = -1, @SerialName("appointmentTime") var appointmentTime: Int? = -1,
     @SerialName("day") var appointmentDay: Int? = -1, @SerialName("month") var appointmentMonth: Int? = -1,
     @SerialName("year") var appointmentYear: Int? = -1, @SerialName("therapist_info") val therapistInfo: TherapistInfo? = null,
     @SerialName("service_type") var serviceTypeItem: ServiceTypeItem? = null, @SerialName("service") var services: Services? = null, @SerialName("time") val platformTime: PlatformTime = PlatformTime(),
