+++ composeApp/src/commonMain/kotlin/domain/Models/TherapistInfo.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable @Parcelize
-data class TherapistInfo(@SerialName("id") val id: Long? = -1, @SerialName("therapist_id") val therapistId: Int? = -1,
+data class TherapistInfo(@SerialName("id") val id: Long? = -1, @SerialName("therapist_id") val therapistId: Long? = -1,
                          @SerialName("isAvailable") val isAvailable: Boolean? = false,
                          @SerialName("isMobileServicesAvailable") val isMobileServiceAvailable: Boolean? = false,
                          @SerialName("profile_info") val profileInfo: User? = null,
