+++ composeApp/src/commonMain/kotlin/domain/bookings/BookingRepository.kt
@@ -6,9 +6,9 @@ import domain.Models.ServerResponse
 import domain.Models.ServiceTherapistsResponse
 
 interface BookingRepository {
-    suspend fun getServiceTherapist(serviceTypeId: Int, vendorId: Long): Single<ServiceTherapistsResponse>
+    suspend fun getServiceTherapist(serviceTypeId: Long, vendorId: Long): Single<ServiceTherapistsResponse>
 
-    suspend fun createPendingBookingAppointment(userId: Long, vendorId: Long, serviceId: Int, serviceTypeId: Int, therapistId: Int,
+    suspend fun createPendingBookingAppointment(userId: Long, vendorId: Long, serviceId: Long, serviceTypeId: Long, therapistId: Long,
                                                 appointmentTime: Int, day: Int, month: Int, year: Int, serviceLocation: String, serviceStatus: String,
                                                 paymentAmount: Double, paymentMethod: String, bookingStatus: String): Single<PendingBookingAppointmentResponse>
     suspend fun createAppointment(userId: Long, vendorId: Long, paymentAmount: Double, paymentMethod: String, bookingStatus: String, day: Int, month: Int, year: Int) : Single<ServerResponse>
